+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.10.0+beta
+
+**Breaking change info:**
+
+This deprecates `upcode flutter:fad` but it doesn't remove it yet. Make sure to update to `upcode fad upload`. 
+
+Updates:
+* implement firebase app distribution in dart. This allows us not to install the node-js firebase-tools package.
+* add `update fad deleteOldReleases` where you can delete old releases on firebase app distribution.
+* update dependencies
+
+
 ## 0.9.3
 
 Reverts 0.9.1 from using the current fvm installation since it doesn't work with CI/CD.